empty string returns correctly
A full block of padding correctly decrypts to `b''` . This empty byte string is set to `dpt` in `admin.py`. A padding error is returned when `not dpt` is true. Unfortunately, `not b''` returns true so the server accidentally sends a  padding error when the full padding block is decrypted correctly.